# swagger.yaml
openapi: 3.0.0
info:
  title: DAPP Votaciones - OpenAPI 3.0
  description: 'API creada para el control de votaciones de la COAC '
  termsOfService: http://swagger.io/terms/
  contact:
    email: ayrtonfidel9@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.1
schemes:
  - http
  - https
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /socios:
    get:
      security:
        - BearerAuth: []
      tags:
        - socios
      summary: 'Traer todos los socios'
      description: 'Endpoint para traer todos los socios de la base de datos, tabla SOCIOS'
      operationId: CU-ND-005
      responses:
        '200':
          description: Datos encontrados correctamente
        '400':
          description: Operacion no autorizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: Datos no encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
  /socios/innerjoin/cuentas:
    get:
      security:
        - BearerAuth: []
      tags:
        - socios
      summary: 'Traer todos los socios'
      description: 'Endpoint para traer todos los socios Y SUS CUENTAS de la base de datos, tabla SOCIOS'
      operationId: CU-ND-012
      responses:
        '200':
          description: Datos encontrados correctamente
        '400':
          description: Operacion no autorizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: Datos no encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
  /socios/registrar:
    post:
      security:
        - BearerAuth: []
      tags:
        - socios
      summary: Registrar un socio de forma individual
      description: 'Registrar un socio individualmente en la base de datos'
      operationId: CU-05
      requestBody:
        description: 'Ingresar un socio'
        content:
          multipart/form-data:
            description: 'Ingresar datos de un socio'
            required: true
            schema:
              $ref: '#/components/schemas/Socios'
      responses:
        '200':
          description: Socio agregado correctamente
        '401':
          $ref: '#/components/schemas/Unauthorized'
  /socios/carga-masiva:
    post:
      security:
        - BearerAuth: []
      tags:
        - socios
      summary: Registrar un socio de forma masiva
      description: 'Registrar socios masivamente en la base de datos'
      operationId: CU-050
      requestBody:
        description: 'Ingresar socios masivamente'
        content:
          multipart/form-data:
            description: 'Ingresar informacion de carga de datos masiva'
            required: true
            schema:
              $ref: '#/components/schemas/SociosMasivos'
      responses:
        '200':
          description: Socio agregado correctamente
        '401':
          $ref: '#/components/schemas/Unauthorized'
  /socios/{idSocio}:
    get:
      security:
        - BearerAuth: []
      tags:
        - socios
      summary: 'Buscar un socio por id'
      description: 'Endpoint para buscar una socio por id en la base de datos, tabla SOCIOS'
      operationId: CU-07
      parameters:
        - in: path
          name: idSocio
          schema:
            type: integer
          required:
          description: El ID del socio debe ser un valor numerico para poder ejecutar la búsqueda correctamente
      responses:
        '200':
          description: Datos encontrados correctamente
        '400':
          description: Operacion no autorizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: Datos no encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
  /socios/existbyPhone/{number}:
    get:
      tags:
        - socios
      summary: 'Verificar la existencia del socio por su numero celular'
      description: 'Endpoint para verificar la existencia del socio, tabla SOCIOS'
      operationId: CU-ND-001
      parameters:
        - in: path
          name: number
          schema:
            type: string
          required:
          description: El numero del socio debe ser un valor numerico para poder ejecutar la búsqueda correctamente
      responses:
        '200':
          description: Datos encontrados correctamente
        '400':
          description: Operacion no autorizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: Datos no encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
  /socios/update/{idSocio}:
    put:
      security:
        - BearerAuth: []
      tags:
        - socios
      summary: 'Actualizar los campos de un socio'
      description: 'Endpoint para actualizar los datos de un socio por el ID, tabla SOCIOS'
      operationId: CU-06
      parameters:
        - in: path
          name: idSocio
          schema:
            type: integer
          required: true
          description: El id debe ser un valor numerico, para  actualizar los datos correctamente
      requestBody:
        description: 'Aztualizar los datos de un socio'
        content:
          multipart/form-data:
            description: 'Ingresar datos de un socio'
            required: true
            schema:
              $ref: '#/components/schemas/Socios'
      responses:
        '200':
          description: Socio actualizado correctamente
        '401':
          $ref: '#/components/schemas/Unauthorized'
  /socios/delete/{idSocio}:
    delete:
      security:
        - BearerAuth: []
      tags:
        - socios
      summary: 'Eliminar un socio por ID'
      description: 'Endpoint para eliminar un socio de la tabla SOCIOS'
      operationId: CU-08
      parameters: 
        - in: path
          name: idSocio
          schema:
            type: integer
          required: true
          description: El ID debe ser un valor numérico para eliminar de forma satisfactoria los datos del socio
      responses:
        '200':
          description: Datos eliminados correctamente
        '400':
          description: Operacion no autorizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: Datos no encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
...
components:
  schemas:
...
    SociosMasivos:
      type: object
      properties:
        idAgencia:
          type: number
          example: 1
        datos:
          type: string
          format: binary
    Socios:
      type: object
      properties:
        nombres:
          type: string
          example: Jose Luis
        apellidos:
          type: string
          example: Almeida Ramirez
        cedula:
          type: string,
          example: '0606252138'
        codigo:
          type: integer
          example: 1091
        imagen:
          type: string
          format: binary
        estado:
          type: boolean
          example: true
        email:
          type: string
          example: joselui@mail.com
        celular:
          type: string
          example: '0981509821'
        idAgencia:
          type: integer
          example: 1
...
    NotFound:
      type: object
      properties:
        message:
          type: string
          example: Recurso no encontrado en la base de datos
    Unauthorized:
      type: object
      properties:
        message:
          type: string
          example: Su rol le impide realizar la presente operación
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT 
...